add instance modify and new function

+++ src/views/rms/instance/InsManage.vue
@@ -25,10 +25,11 @@
             <el-table-column prop="mode" align="center" label="连接信息">
                 <template slot-scope="scope">
                     <el-tag v-if="scope.row.mode==='standalone'">{{scope.row.ip+':'+scope.row.port}}</el-tag>
-                    <el-tag v-if="scope.row.mode!=='standalone'">Cluster</el-tag>
+                    <el-tag v-if="scope.row.mode===undefined">Unknown</el-tag>
+                    <el-tag v-if="scope.row.mode==='cluster'">Cluster</el-tag>
                 </template>
             </el-table-column>
-            <el-table-column prop="state" width="80px" align="center" label="状态">
+            <el-table-column prop="state" width="70px" align="center" label="状态">
                 <template slot-scope="scope">
                     <el-tag :type="scope.row.state ? 'success' : 'danger'">{{scope.row.state?'在线':'断开'}}
                     </el-tag>
@@ -37,18 +38,18 @@
             <el-table-column prop="uptime_in_seconds" align="center" width="160px" label="运行时长">
                 <template slot-scope="scope">{{parseTime(scope.row.uptime_in_seconds)}}</template>
             </el-table-column>
-            <el-table-column prop="mode" align="center" width="80px" label="模式"/>
+            <el-table-column prop="mode" align="center" width="100px" label="模式"/>
             <el-table-column prop="role" align="center" width="80px" label="角色"/>
             <el-table-column prop="os" align="center" width="80px" label="操作系统"/>
-            <el-table-column prop="version" align="center" width="80px" label="版本"/>
+            <el-table-column prop="version" align="center" width="70px" label="版本"/>
             <el-table-column prop="arch_bits" align="center" width="50px" label="架构">
                 <template slot-scope="scope">{{'x'+scope.row.arch_bits}}</template>
             </el-table-column>
-            <el-table-column prop="user" align="center" label="用户" width="50px" v-if="loadColStatus"/>
-            <el-table-column prop="create_time" align="center" width="190px" label="创建时间">
+            <el-table-column prop="user" align="center" label="用户" width="80px" v-if="loadColStatus"/>
+            <el-table-column prop="ct" align="center" width="190px" label="创建时间">
                 <template slot-scope="scope">
                     <i class="el-icon-time"/>
-                    <span style="margin-left: 10px">{{  dateFormat("YYYY-mm-dd HH:MM:SS",new Date(scope.row.create_time)) }}</span>
+                    <span style="margin-left: 10px">{{  dateFormat("YYYY-mm-dd HH:MM:SS",new Date(scope.row.ct)) }}</span>
                 </template>
             </el-table-column>
             <el-table-column align="center" label="操作">
@@ -62,7 +63,7 @@
                     <el-button v-if="rights('INS_UPDATE_INFO')" type="text" size="small"
                                @click="editName(scope.row)">修改名称
                     </el-button>
-                    <el-button v-if="rights('INS_REFRESH_META')" type="text" size="small"
+                    <el-button v-if="rights('INS_REFRESH_META')" type="text" size="small" :disabled="!scope.row.state"
                                @click="refreshInfo(scope.row)">刷新
                     </el-button>
                     <el-button v-if="rights('INS_DEL')" type="text" size="small" @click="deleteInstance(scope.row)">删除
@@ -74,7 +75,14 @@
 </template>
 
 <script>
-    import {deleteIns, getAll, getByUser, modifyState, refreshMeta} from "../../../api/redismanage/redis_ins";
+    import {
+        deleteIns,
+        getAll,
+        getByUser,
+        modifyName,
+        modifyState,
+        refreshMeta
+    } from "../../../api/redismanage/redis_ins";
     import {dateFormat, parseSec} from "../../../utils/utils";
     import InstanceForm from "./InstanceForm";
 
@@ -87,7 +95,7 @@
                 identify: this.$store.getters.identify,
                 query: {
                     blurry: ''
-                }, tableData: [],
+                }
             }
         }, created() {
             this.$nextTick(() => {
@@ -170,8 +178,40 @@
                 tmpForm.id = row.id;
                 _this.form = tmpForm;
                 _this.dialog = true;
-            }, editName() {
-
+            }, editName(row) {
+                this.$prompt('请输入新名称', '提示', {
+                    confirmButtonText: '确定',
+                    cancelButtonText: '取消',
+                    inputErrorMessage: '名称格式不正确'
+                }).then(({value}) => {
+                    let form = {
+                        id: row.id,
+                        name: value
+                    };
+                    modifyName(form).then((resp) => {
+                        if (resp.data.status === 200 && resp.data.code === 'OK') {
+                            this.$message({
+                                type: 'success',
+                                message: '名称修改成功!'
+                            });
+                            row.instance_name = resp.data.message.instance_name;
+                        } else {
+                            this.$message({
+                                type: 'warning',
+                                message: resp.data.message
+                            });
+                        }
+                    }).catch(() => {
+                        this.$message.error({
+                            message: "名称修改错误!"
+                        });
+                    })
+                }).catch(() => {
+                    this.$message({
+                        type: 'info',
+                        message: '取消修改'
+                    });
+                });
             }, changeState(row) {
                 let dto = {
                     id: row.id,
@@ -219,25 +259,48 @@
                     });
                 })
             }, deleteInstance(row) {
-                deleteIns(row.id).then((resp) => {
-                    if (resp.data.status === 200 && resp.data.code === "OK") {
-                        this.$message({
-                            type: 'success',
-                            message: '删除成功!'
+                this.$confirm('将删除此实例并清空该实例所有数据, 是否继续?', '警告', {
+                    confirmButtonText: '确定',
+                    cancelButtonText: '取消',
+                    type: 'warning'
+                }).then(() => {
+                    this.lockLoading();
+                    deleteIns(row.id).then((resp) => {
+                        if (resp.data.status === 200 && resp.data.code === "OK") {
+                            this.$message({
+                                type: 'success',
+                                message: '删除成功!'
+                            });
+                        } else {
+                            this.$message({
+                                type: 'warning',
+                                message: '删除失败:' + resp.data.message
+                            });
+                        }
+                        this.loadData();
+                        this.unlockLoading();
+                    }).catch(() => {
+                        this.loadData();
+                        this.unlockLoading();
+                        this.$message.error({
+                            message: "删除错误"
                         });
-                    } else {
-                        this.$message({
-                            type: 'warning',
-                            message: '删除失败:' + resp.data.message
-                        });
-                    }
-                    this.loadData();
+                    })
                 }).catch(() => {
-                    this.loadData();
-                    this.$message.error({
-                        message: "删除错误"
+                    this.$message({
+                        type: 'info',
+                        message: '已取消删除'
                     });
-                })
+                });
+            }, lockLoading() {
+                this.loadBg = this.$loading({
+                    lock: true,
+                    text: 'Loading',
+                    spinner: 'el-icon-loading',
+                    background: 'rgba(0, 0, 0, 0.7)'
+                });
+            }, unlockLoading() {
+                this.loadBg.close();
             }
         }
     }
